refactor: use index.ts for re-export component
