+++ src/components/models/memo-dialog/MemoDialogContentsConstructor/index.ts
@@ -1,0 +1,1 @@
+export { MemoDialogContentsConstructor } from './MemoDialogContentsConstructor'
